requirements.txt
python-telegram-bot==13.15
python-dotenv==1.1.0
urllib3==1.26.8
redis-6.2.0